Add debug files
